Makefile: Poplar-related changes
Disable the handling as errors use of __DATE__ and __TIME__ in the
code.  Some of the Poplar code currently uses one or both of those
in definition of a version string.  Until we fix that, we need to
disable this.

One other change is that symbolic links are followed when searching
for files for "make clean" and "make distclean".  This too can go
away now; we don't use symbolic links in the kernel tree any more.

Signed-off-by: Alex Elder <elder@linaro.org>