drivers: a large mix of remaining code
This patch combines a bunch of code that was too difficult
to sort out into smaller features in the time I had.

It could be split into pieces by someone else, or reviewed
as a whole, and perhaps some features could be pulled out
of it to make rebases easier.

Signed-off-by: Alex Elder <elder@linaro.org>