drivers/msp/include: delete a lot of definitions
There are a lot of definitions of under "drivers/msp/include" that
are never used.  The ones deleted here are just the result of sort
of randomly poking around in a few files, and deleting things that
are never used.  There could easily be a lot more patches like this.

Signed-off-by: Alex Elder <elder@linaro.org>