drivers: get rid of more __cplusplus #ifdef's
We currently aren't providing a formal user space API for the
HiSilicon code.  It's likely that header files found under
drivers/sdk are at least part of that, but at some point we need to
do a better job of really exposing the interface shared between the
kernel and user space.

Until we have that, there's no reason to protect against compilation
by a C++ compiler, certainly in header files not clearly meant to be
exported there.

So get rid of these "ifdef __cpluplus" blocks at the beginning and
end of C source files that contain them.

Signed-off-by: Alex Elder <elder@linaro.org>